requirements.txt
Flask==2.0.1
line-bot-sdk==1.8.0